+++ packages/webpack/__e2e-fixtures__/sass-extract-text/webpack.config.js
@@ -1,10 +1,11 @@
-const { createConfig, entryPoint, match, setOutput } = require('../../index')
+const { createConfig, entryPoint, match, setMode, setOutput } = require('../../index')
 
 const path = require('path')
 const extractText = require('@webpack-blocks/extract-text')
 const sass = require('@webpack-blocks/sass')
 
 module.exports = createConfig([
+  setMode('development'),
   entryPoint(
     path.join(__dirname, 'app.js')
   ),