+++ packages/dev-server/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @webpack-blocks/dev-server - Changelog
 
+## Next Release
+
+- Removed `NamedModulesPlugin` because webpack 4 automatically adds it in the `development` mode ([#261](https://github.com/andywer/webpack-blocks/pull/261))
+
 ## 1.0.0
 
 - Removed `proxy()` ([#251](https://github.com/andywer/webpack-blocks/pull/251))