+++ packages/uglify/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @webpack-blocks/uglify - Changelog
 
+## Next Release
+
+- Use the new `optimization` API. ([#255](https://github.com/andywer/webpack-blocks/issues/255))
+
 ## 1.1.0
 
 - Use `uglifyjs-webpack-plugin` v1.0 (not the beta) ([#231](https://github.com/andywer/webpack-blocks/issues/231))